demo/requirements.txt
# Flask Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Machine Learning
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1

# Utilities
python-dateutil==2.8.2
werkzeug==2.3.7

# Optional: For production deployment
gunicorn==21.2.0
